Improve INSTALL.md; remove apt update from install-apt.sh (#239)
* Remove apt update from install-apt.sh, and instead indicate in
installation manual that packages should be updated at the discretion of
the user.
* Cleaned up installation instructions

This was motivated by attempting to set up Bystro on a new Ubuntu
server. I realized that users will not want Bystro to force package
updates, and that the installation of the Python libraries is not
spelled out in INSTALL.md

Future PRs will strip out sudo from the install scripts, this is a
gentle first step.