[annotator] Use NA as the emptyField (#345)
* Use NA as the empty field value everywhere

This ensures that pandas and R read missing values as missing. Better to
go with the tide on these things: while "!" is more space efficient, the
cons for downstream processing are large: you need to override default
missing value in data frame libraries.


Almost all of the changes are in "*.yml". Ignore those. It's 99%
formatting changes, the only other difference is using "--emptyField NA"
instead of "--emptyField '!'" and the hg19.clean.yml is updated to
reflect the newest database build. Ignore these and assume they are
right.