[annotator] Add no_build option which allows tracks to be skipped (#346)
* Add no_build option which allows tracks to be skipped
* Cleans up unnecessary DDP imports

This is necessary, because in some cases, like join tracks, we may not
want to build the track referenced in the join definition. This comes up
with clinvar, which defines large variants and small in the tsv output,
but not in the vcf output. So we define a joint track "refSeq.clinvar"
to show those large structural variants that intersect with refSeq, but
do not want to build the small variants in the actual clinvar track,
because those are inexact in position (we instead now have an exact
match clinvar2).