Bugfix marginal likelihood supervision + SVAE (#464)
This combines two tasks into a single PR. We fix one of the formulas in
the marginal supervision model, which can be seen in lines 166-168 of
gf_marginal, where the wrong loss function for the predictive likelihood
is used. We also implemented a linear Supervised Variational
Autoencoder, which offers no practical advantages, but is an important
comparison method in our paper.