Covariance Module Improvements (#465)
We are now realizing that fitting covariance matrices is a critical task
and while there is extensive work in the statistics community on
improving finite-sample properties of estimation, very little has
migrated to other fields. Part of this is likely due to scattered
implementations across a variety of github repositories. We are
implementing a large quantity of these estimators ourselves with a
common format similar to sklearn. Here I implement two shrinkage
estimators by Ledoit and Wolfe.