[annotation] Add isCanonical transcript annotation (#478)
* Creates isCanonical transcript feature, which will be outputted as
'TRUE' or 'FALSE' depending on whether there is a matching UCSC
knownCanonical feature


We output 'TRUE', 'FALSE' here instead of 1/0, because pandas, polars
(the most popular data science libraries in python), R standard
dataframe library, and opensearch will interpret this as the boolean
True, or False. We also use TRUE, FALSE for discordant for similar
reasons. In the case of refSeq this isn't as useful because we often
have multiple transcripts, so I'm open to going back to using 0/1.