Enhance Covariance Matrix Estimation with Symmetry Enforcement and Warnings (#480)
This pull request introduces an improvement in the covariance matrix
estimation process by ensuring that the matrix remains symmetric
throughout the computation. We have implemented an auxiliary function,
`symmetrize_and_warn`, that adjusts the matrix by setting `S_new = (S +
S.T) / 2`. This function is designed to enforce symmetry and alert the
user when significant adjustments are made to the matrix.

### Key changes include:
- Addition of the `symmetrize_and_warn` function which symmetrizes the
covariance matrix and generates warnings if the modifications are
substantial.
- Integration of this function into the existing matrix estimation
workflow to automatically ensure the symmetry of the output.

This enhancement helps prevent errors in subsequent analyses that rely
on the symmetry of the covariance matrix, thereby increasing the
robustness of our statistical computations.