[infrastructure] make canopy direct dependency until SomaLogic makes pypi package (#499)
* Fixes release creation, which was failing because we rely on Canopy,
which does not have a pypi package. To solve this we update our
instructions to inform users that they for now need to install canopy
* fixes perltidy error, which for some reason was only see on release
* fixes how liftover is imported, so we do not query the UCSC server
unless necessary
* updates liftover dependency, so that it can use alternate UCSC mirrors
* adds missing ancestry api test

This solution is temporary; once SomaLogic makes a pypi package, we will
depend on that instead: https://github.com/SomaLogic/Canopy/issues/6. If
they choose not to, we will publish our own.