[installation] Pin OSQP version to ensure platform compatibility (#560)
* When installing with --pre, on Thomas' Mac, when OSQP >= 0.6.2 is
requested by cvxpy, osqp==1.0.0b1 is installed instead, and that version
cannot install on Mac OS 14. So we now pin osqp==0.6.3 to satisfy
cvxpy's requirements without allowing the pip to install an incompatible
beta release of osqp. We pin to osqp==0.6.5, which is known to work on
Mac, and is a recent release (May 2024).
* Updates cvxpy to 1.5.2, which fixes some small bugs and adds numpy 2.0
compatibility

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Introduced a new optimization library (`osqp`), enhancing performance
for quadratic programming tasks.
- **Updates**
- Upgraded the existing `cvxpy` library to version 1.5.2, which may
include various improvements and bug fixes.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->